Added code that makes sure all of the books in the library are listed
The code loops, following the "next" links of the feed until there are
no more "next" links.

It was necessary to make the Metadata.timestamp a real timestamp to stop
the debug trace from failing to output it.  This resulted in the updated
colum of the OPDS book list to become empty.