Emit model change signal between each OPDS chunk read.
The OPDS feed from calibre is read in chunks of 30, and the
reading/parsing of a chunk consumes quite a bit of clock
time.

Tried emitting a changed signal from the model between each
chunk, in the hope that this would visually update the book
list.  Unfortunately there still are no visible changes in
the list until the read is complete