Set the Metadata timestamp correctly using the calibre REST API.
The updated values returned in the OPDS feeds from calibre, are the same
for all books, and unrelated to the books themselves.  What's returned
is the last modified date of the calibre database.

The calibre REST API offers the possibility of retrieving the actual
timestamp of the books, so the code now detects if the server is
calibre, and if it is, uses the calibre REST API to get the timestamps.