Made the "Fix timestamps of selection button" work for books with more than one author.
Workaround for strange Cache.find_identical_books() behaviour: the
"superset" works for a single author but not with all authors in
the case of one or more authors.

The workaround for books with more than one author: match the authors
one by one and make a union of the matched book ids.